requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
frozenlist==1.4.1
idna==3.7
lxml==5.2.2
multidict==6.0.5
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
tzdata==2024.1
yarl==1.9.4
